Return an error rather than panicking on a bad request.
It's considered much more idiomatic to return errors rather than
panicking if something goes wrong. It might actually be _dangerous_
to panic here... the expectation is that if there's a possiblity
for an error, an error will be returned. If consumers don't see that
an error is returned, they may simply assume `nil` is returned
on an invalid request, and applications will burn...